refactor: Updated log messages and fixed namespaces
