+++ tests/screeningDataServicesTests/markParticipantAsEligibleTests/markParticipantAsEligibleTests.cs
@@ -8,7 +8,7 @@ using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.Extensions.Logging;
 using Model;
 using Moq;
-using screeningDataServices;
+using NHS.CohortManager.Tests.TestUtils;
 
 [TestClass]
 public class MarkParticipantAsEligibleTests
